pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-separation"
version = "0.1.0"
description = "Modern speech separation system for cocktail party problem"
authors = [{name = "AI Research Team"}]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "soundfile>=0.12.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "tensorboard>=2.13.0",
    "asteroid>=0.6.0",
    "pesq>=0.0.4",
    "pystoi>=0.3.3",
    "streamlit>=1.25.0",
    "gradio>=3.40.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.280",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.3.0",
]
tracking = [
    "wandb>=0.15.0",
    "mlflow>=2.5.0",
]
serving = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "W503", "COM812", "ISC001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
